remove node and sort string
