+++ src/sort.ts
@@ -1,4 +1,4 @@
-import { Node } from "./linkedlist";
+import { Node } from ".";
 
 /**
  * 
@@ -26,26 +26,42 @@ export default class Sort<T> {
         return slow;
     }
 
-    sortedMerge(a: Node<T> | null | undefined, b: Node<T> | null | undefined) {
+    sortedMerge(a: Node<T> | null | undefined, b: Node<T> | null | undefined, order: number) {
         var result: Node<T> | null | undefined = null;
         /* Base cases */
         if (a == null)
             return b;
         if (b == null)
             return a;
-
-        /* Pick either a or b, and recur */
-        if (parseFloat("" + a?.val) <= parseFloat("" + b?.val)) {
-            result = a;
-            result.next = this.sortedMerge(a?.next, b);
-        } else {
-            result = b;
-            result.next = this.sortedMerge(a, b?.next);
+        console.log("order", order);
+        switch (order) {
+            case -1: {
+                if (<T>a?.val <= <T>b?.val) {
+                    result = b;
+                    result.next = this.sortedMerge(a, b?.next, order);
+                } else {
+                    result = a;
+                    result.next = this.sortedMerge(a?.next, b, order);
+                }
+                break;
+            }
+            default: {
+                if (<T>a?.val <= <T>b?.val) {
+                    result = a;
+                    result.next = this.sortedMerge(a?.next, b, order);
+                } else {
+                    result = b;
+                    result.next = this.sortedMerge(a, b?.next, order);
+                }
+            }
         }
+        /* Pick either a or b, and recur */
+        // if (parseFloat("" + a?.val) <= parseFloat("" + b?.val)) {
+
         return result;
     }
 
-    mergeSort(h: Node<T> | null | undefined) {
+    mergeSort(h: Node<T> | null | undefined, order: number = 1) {
         // Base case : if head is null
         if (h == null || h.next == null) {
             return h;
@@ -60,15 +76,13 @@ export default class Sort<T> {
             middle.next = null;
 
         // Apply mergeSort on left list
-        var left: Node<T> | null | undefined = this.mergeSort(h);
+        var left: Node<T> | null | undefined = this.mergeSort(h, order);
 
         // Apply mergeSort on right list
-        var right: Node<T> | null | undefined = this.mergeSort(nextofmiddle);
+        var right: Node<T> | null | undefined = this.mergeSort(nextofmiddle, order);
 
         // Merge the left and right lists
-        var sortedlist = this.sortedMerge(left, right);
+        var sortedlist = this.sortedMerge(left, right, order);
         return sortedlist;
     }
-
-
 }
